category permission created successfully

+++ app/Http/Controllers/RoleController.php
@@ -18,7 +18,7 @@ class RoleController extends Controller
      */
     function __construct()
     {
-        $this->middleware('permission:role-list|role-create|role-edit|role-delete', ['only' => ['index','store']]);
+        $this->middleware('permission:role-list|role-create|role-edit|role-delete', ['only' => ['index','show']]);
         $this->middleware('permission:role-create', ['only' => ['create','store']]);
         $this->middleware('permission:role-edit', ['only' => ['edit','update']]);
         $this->middleware('permission:role-delete', ['only' => ['destroy']]);
@@ -130,6 +130,7 @@ class RoleController extends Controller
      */
     public function destroy($id)
     {
+        if ($id == 1) abort(403);
         DB::table("roles")->where('id',$id)->delete();
         return redirect()->route('admin.roles.index')
             ->with('success','Role deleted successfully');
